+++ jeap-messaging-sequential-inbox/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/jpa/SpringDataJpaSequenceInstanceRepository.java
@@ -3,6 +3,7 @@ package ch.admin.bit.jeap.messaging.sequentialinbox.jpa;
 import ch.admin.bit.jeap.messaging.sequentialinbox.persistence.SequenceInstance;
 import ch.admin.bit.jeap.messaging.sequentialinbox.persistence.SequenceInstanceState;
 import jakarta.persistence.LockModeType;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
@@ -56,5 +57,14 @@ interface SpringDataJpaSequenceInstanceRepository extends JpaRepository<Sequence
     @Modifying
     @Query("DELETE FROM SequenceInstance si WHERE si.id = :id AND si.state <> :state")
     int deleteByIdAndStateNot(@Param("id") long sequenceInstanceId, @Param("state") SequenceInstanceState state);
+
+    List<SequenceInstance> findAllByPendingActionIsNotNull();
+
+    @Query(nativeQuery = true, value = "SELECT * from sequence_instance where created_at + (0.75 * EXTRACT(EPOCH FROM retain_until - created_at)) * INTERVAL '1 second' < now()")
+    Page<SequenceInstance> findAllWithRetentionPeriodElapsed75Percent(Pageable pageable);
+
+    Page<SequenceInstance> findAllByRetainUntilBefore(ZonedDateTime retainUntilBefore, Pageable pageable);
+
+    Optional<SequenceInstance> findByNameAndContextId(String name, String contextId);
 
 }