+++ CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [11.2.0] - 2025-11-21
+### Changed
+- add new pending action to sequence instance and sequenced message in order to support devOps operations
+
 ## [11.1.0] - 2025-11-14
 ### Changed
 - update jeap-messaging from 10.0.1 to 10.1.0